// src/types/JunkShape.ts
export enum JunkShape {
  // standard shapes
  RECT_2X3 = 'RECT_2X3',
  RECT_3X2 = 'RECT_3X2',
  RECT_3X3 = 'RECT_3X3',
  RECT_4X2 = 'RECT_4X2',
  BIG_L = 'BIG_L',
  BIG_L_REV = 'BIG_L_REV',
  // special shapes
  RECT_1X1 = 'RECT_1X1',
  RECT_2X1 = 'RECT_2X1',
  RECT_2X2 = 'RECT_2X2',
  RECT_4X4 = 'RECT_4X4',
  RECT_10X1 = 'RECT_10X1',
}